break menu into its own component.
